created project_explorer_models.py and moved func.py in the application root directory
